New regression test for root finding
As detailed in issue #65 there is a problem with our polynomial interpolation /
root-finding process if the levels are very close together (GFDL ran into an
issue when levels were 1e-10 m thick). This commit adds a new test to try to
reproduce that problem and also see if normalizing the depths before
interpolating would help avoid the problem.