Sign error!
In CVMix KPP, OBL_depth = -root (because depth is positive up and we want the
OBL depth to be a positive value). When I wrote this driver, I included the
negative sign but didn't account for it properly when converting the normalized
root to true depth. For simplicity, I have just removed the sign altogether.