Add subtest to call conv mixing from KPP test
Added a call to cvmix_coeffs_conv() from cvmix_kpp_driver() because a user was
having trouble with calling this routine with specific ifortran compile flags
when compiling with intel 19.0.4. I also updated bld/CompileFlags.mak to try to
reproduce the error, but 19.0.5 did not reproduce the runtime error
encountered.